requirements.txt
# PROJECT SPECIFIC REQUIREMENTS
cysignals
# If pyfasttext setup fails, install Cython manually first per pip install Cython
Cython
ftfy
jupyterlab
markdown==3.1.1
matplotlib==3.1.0
nltk==3.4.4
numpy==1.16.4
pandas==0.24.2
praw==6.3.1
pytest==4.6.3
ruamel.yaml==0.15.94
seaborn==0.9.0
spacy==2.1.4
sqlitedict
textstat==0.5.6
sentencepiece==0.1.85
# CUDA (10):
# pip3 install torch==1.3.1
# CPU only:
# pip3 install torch==1.3.1+cpu -f https://download.pytorch.org/whl/torch_stable.html
transformers==2.2.2
tables==3.5.2
tqdm==4.32.2
voussoirkit==0.0.30
pyfasttext
